Fixed __init__ and moved memoizers back again
