+++ src/qttools/boundary_conditions/__init__.py
@@ -1,6 +1,1 @@
-# Copyright (c) 2025 ETH Zurich and the authors of the qttools package.
-
-from qttools.boundary_conditions.lyapunov import LyapunovSolver, LyapunovSystem
-from qttools.boundary_conditions.obc import OBCSolver, OBCSystem
-
-__all__ = ["LyapunovSystem", "LyapunovSolver", "OBCSystem", "OBCSolver"]
+# Copyright (c) 2024-2026 ETH Zurich and the authors of the qttools package.
