--- src/qttools/boundary_conditions/lyapunov_solver/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2024-2026 ETH Zurich and the authors of the qttools package.
-
-from qttools.boundary_conditions.lyapunov_solver.doubling import Doubling
-from qttools.boundary_conditions.lyapunov_solver.spectral import Spectral
-
-__all__ = ["Doubling", "Spectral"]
