+++ src/qttools/boundary_conditions/obc/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2024-2025 ETH Zurich and the authors of the qttools package.
+
+from qttools.boundary_conditions.obc.obc import OBCSolver, OBCSystem
+from qttools.boundary_conditions.obc.sancho_rubio import SanchoRubio
+from qttools.boundary_conditions.obc.spectral import Spectral
+
+__all__ = ["SanchoRubio", "Spectral", "OBCSystem", "OBCSolver"]
