--- src/qttools/boundary_conditions/obc_solver/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (c) 2024-2026 ETH Zurich and the authors of the qttools package.
-
-from qttools.obc.obc import OBCMemoizer, OBCSolver
-from qttools.obc.sancho_rubio import SanchoRubio
-from qttools.obc.spectral import Spectral
-
-__all__ = ["OBCSolver", "OBCMemoizer", "SanchoRubio", "Spectral"]
